endpoint/requirements.txt
simplejson
pandas
flask
google-cloud-storage
pystan==2.19.1.1
fbprophet